x86: make ASID field global
Make processors allocate the same ASID to the same guest. This is a
prerequisite for having guests use the AMD INVLPGB instruction and for
supporting encrypted VMs in Xen.

Set the ASID for every L1 guest's CPUs equal to the ID of the
guest. Use a separate ASID for L2 guest CPUs (bit 16) to avoid
flushing the TLB on every L1/L2 guest transition. Currently, dom_id is
15 bits wide, leaving at least one empty bit for nested guests for a
16-bit (Intel VPID) or wider ASID value. Crash the kernel if there is
ever a patch that increases DOMID_MASK.

Remove ASID core generation logic.

RFC because:

- I tried to leverage the existing SVM TLB Control definitions for
  simplicity.

- HVM_ENTRY_TLB_* might not be the best prefix.

- I moved need_flush from the hvm_asid_data per-CPU structure to
  hvm_vcpu_asid to enable deferring the flush for when vCPU is running
  in guest mode and vice-versa.

Suggested-by: Andrew Cooper <andrew.cooper3@citrix.com>
Signed-off-by: Jane Malalane <jane.malalane@citrix.com>
---
CC: Jan Beulich <jbeulich@suse.com>
CC: Andrew Cooper <andrew.cooper3@citrix.com>
CC: "Roger Pau Monné" <roger.pau@citrix.com>
CC: Wei Liu <wl@xen.org>
CC: Jun Nakajima <jun.nakajima@intel.com>
CC: Kevin Tian <kevin.tian@intel.com>
CC: George Dunlap <george.dunlap@citrix.com>