x86/entry: Use linkage.h helpers to add tags to symbols
Adding a start and end type-specific label as introduced by the
linkage.h code defines a size and type for the symbol which is then
referenced in the symbol table, unlike previous behaviour where the
symbol size and type were unknown.

This is a step toward fixing livepatching within .text.entry as
defining a type and size for a symbol aids the livepatch build tool in
detecting changes made within that section.

---
CC: Jan Beulich <jbeulich@suse.com>
CC: Andrew Cooper <andrew.cooper3@citrix.com>
CC: "Roger Pau Monné" <roger.pau@citrix.com>
CC: Wei Liu <wl@xen.org>